4.20: authentication enabled for modify blog

+++ osa4/blogilista/controllers/blogs.js
@@ -57,9 +57,17 @@ blogsRouter.put('/:id', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
-    await Blog
-      .findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id)
+      return response.status(401).json({ error: 'invalid or missing token' })
+    const blog = await Blog.findById(request.params.id)
+    if (!blog) return response.status(204).end()
+    if ( blog.user.toString() === decodedToken.id.toString() ) {
+      await Blog
+        .findByIdAndRemove(request.params.id)
+      return response.status(204).end()
+    }
+    return response.status(401).json({ error: 'invalid token' })
   } catch(exception) {
     next(exception)
   }
